Renaming OSMPreparationChain to PrepareOSM. Moving the corresponding files from prepare_data to processing_chain.
