+++ prepare_data/src/main/groovy/org/orbisgis/PrepareData.groovy
@@ -5,7 +5,6 @@ import org.orbisgis.common.AbstractTablesInitialization
 import org.orbisgis.osm.FormattingForAbstractModel
 import org.orbisgis.common.InputDataFormatting
 import org.orbisgis.osm.OSMGISLayers
-import org.orbisgis.osm.OSMPreparationChain
 import org.orbisgis.processmanager.ProcessManager
 import org.orbisgis.processmanagerapi.IProcessFactory
 import org.slf4j.Logger
@@ -24,5 +23,4 @@ abstract class PrepareData extends Script {
     public static FormattingForAbstractModel = new FormattingForAbstractModel()
     public static InputDataFormatting = new InputDataFormatting()
 
-    public static OSMPreparationChain = new OSMPreparationChain()
 }
