+++ processing-chain/src/main/groovy/org/orbisgis/processingchain/ProcessingChain.groovy
@@ -12,6 +12,7 @@ abstract class ProcessingChain extends Script {
     public static IProcessFactory processFactory = ProcessManager.getProcessManager().factory("processing_chain")
 
     public  static PrepareBDTopo prepareBDTopo= new PrepareBDTopo()
+    public  static PrepareOSM prepareOSM = new PrepareOSM()
 
     public  static CreateScalesOfAnalysis createScalesOfAnalysis= new CreateScalesOfAnalysis()
